Put app collection into separate users

+++ lib/data/remote_data_sources/remote_data_source.dart
@@ -1,11 +1,20 @@
 import 'dart:async';
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:letsparty/domain/models/theme_model.dart';
 
 class RemoteDataSource {
   Stream<QuerySnapshot<Map<String, dynamic>>> getItemsStream() {
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
     try {
-      final stream = FirebaseFirestore.instance.collection('items').snapshots();
+      final stream = FirebaseFirestore.instance
+          .collection('app')
+          .doc(userId)
+          .collection('items')
+          .snapshots();
       return stream;
     } catch (error) {
       throw Exception(error.toString());
@@ -13,7 +22,16 @@ class RemoteDataSource {
   }
 
   Future<void> removeItems({required String id}) {
-    return FirebaseFirestore.instance.collection('items').doc(id).delete();
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
+    return FirebaseFirestore.instance
+        .collection('app')
+        .doc(userId)
+        .collection('items')
+        .doc(id)
+        .delete();
   }
 
   Future<void> updateItems({
@@ -23,9 +41,17 @@ class RemoteDataSource {
     required DateTime date,
     required String time,
   }) async {
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
     try {
-      final update =
-          await FirebaseFirestore.instance.collection('items').doc(id).update({
+      final update = await FirebaseFirestore.instance
+          .collection('app')
+          .doc(userId)
+          .collection('items')
+          .doc(id)
+          .update({
         'city': city,
         'adress': adress,
         'date': date,
@@ -43,7 +69,15 @@ class RemoteDataSource {
     required DateTime date,
     required String time,
   }) {
-    return FirebaseFirestore.instance.collection('items').add(
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
+    return FirebaseFirestore.instance
+        .collection('app')
+        .doc(userId)
+        .collection('items')
+        .add(
       {
         'city': city,
         'adress': adress,
@@ -55,7 +89,15 @@ class RemoteDataSource {
 
   Stream<QuerySnapshot<Map<String, dynamic>>> getMenuStream() {
     try {
-      final stream = FirebaseFirestore.instance.collection('menu').snapshots();
+      final userId = FirebaseAuth.instance.currentUser?.uid;
+      if (userId == null) {
+        throw Exception('User is not logged in');
+      }
+      final stream = FirebaseFirestore.instance
+          .collection('app')
+          .doc(userId)
+          .collection('menu')
+          .snapshots();
       return stream;
     } catch (error) {
       throw Exception(error.toString());
@@ -65,7 +107,15 @@ class RemoteDataSource {
   Future<void> addMenuDocuments({
     required String title,
   }) async {
-    await FirebaseFirestore.instance.collection('menu').add(
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
+    await FirebaseFirestore.instance
+        .collection('app')
+        .doc(userId)
+        .collection('menu')
+        .add(
       {
         'title': title,
       },
@@ -73,13 +123,29 @@ class RemoteDataSource {
   }
 
   Future<void> removeMenu({required String id}) {
-    return FirebaseFirestore.instance.collection('menu').doc(id).delete();
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
+    return FirebaseFirestore.instance
+        .collection('app')
+        .doc(userId)
+        .collection('menu')
+        .doc(id)
+        .delete();
   }
 
   Stream<QuerySnapshot<Map<String, dynamic>>> getThemeStream() {
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
     try {
-      final stream =
-          FirebaseFirestore.instance.collection('themePhotos').snapshots();
+      final stream = FirebaseFirestore.instance
+          .collection('app')
+          .doc(userId)
+          .collection('themePhotos')
+          .snapshots();
       return stream;
     } catch (error) {
       throw Exception(error.toString());
@@ -89,7 +155,15 @@ class RemoteDataSource {
   Future<void> addThemePhoto({
     required String imageUrl,
   }) async {
-    await FirebaseFirestore.instance.collection('themePhotos').add(
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
+    await FirebaseFirestore.instance
+        .collection('app')
+        .doc(userId)
+        .collection('themePhotos')
+        .add(
       {
         'image_url': imageUrl,
       },
@@ -98,7 +172,16 @@ class RemoteDataSource {
 
   Future<ThemeModel> getPhoto(
       {required String id, required String imageUrl}) async {
-    await FirebaseFirestore.instance.collection('themePhotos').doc(id).get();
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
+    await FirebaseFirestore.instance
+        .collection('app')
+        .doc(userId)
+        .collection('themePhotos')
+        .doc(id)
+        .get();
     return ThemeModel(
       id: id,
       imageUrl: imageUrl,
@@ -106,15 +189,29 @@ class RemoteDataSource {
   }
 
   Future<void> removeThemePhoto({required String id}) {
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
     return FirebaseFirestore.instance
+        .collection('app')
+        .doc(userId)
         .collection('themePhotos')
         .doc(id)
         .delete();
   }
 
   Stream<QuerySnapshot<Map<String, dynamic>>> getUserStream() {
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
     try {
-      final stream = FirebaseFirestore.instance.collection('user').snapshots();
+      final stream = FirebaseFirestore.instance
+          .collection('app')
+          .doc(userId)
+          .collection('user')
+          .snapshots();
       return stream;
     } catch (error) {
       throw Exception(error.toString());
@@ -125,7 +222,15 @@ class RemoteDataSource {
     required String name,
     required String photo,
   }) async {
-    await FirebaseFirestore.instance.collection('user').add(
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
+    await FirebaseFirestore.instance
+        .collection('app')
+        .doc(userId)
+        .collection('user')
+        .add(
       {
         'name': name,
         'photo': photo,
@@ -134,13 +239,29 @@ class RemoteDataSource {
   }
 
   Future<void> removeUserItems({required String id}) {
-    return FirebaseFirestore.instance.collection('user').doc(id).delete();
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
+    return FirebaseFirestore.instance
+        .collection('app')
+        .doc(userId)
+        .collection('user')
+        .doc(id)
+        .delete();
   }
 
   Stream<QuerySnapshot<Map<String, dynamic>>> getBudgetStream() {
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
     try {
-      final stream =
-          FirebaseFirestore.instance.collection('finance').snapshots();
+      final stream = FirebaseFirestore.instance
+          .collection('app')
+          .doc(userId)
+          .collection('finance')
+          .snapshots();
       return stream;
     } catch (error) {
       throw Exception(error.toString());
@@ -150,7 +271,15 @@ class RemoteDataSource {
   Future<void> addBudgetDocuments({
     required String data,
   }) async {
-    await FirebaseFirestore.instance.collection('finance').add(
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
+    await FirebaseFirestore.instance
+        .collection('app')
+        .doc(userId)
+        .collection('finance')
+        .add(
       {
         'data': data,
       },
@@ -161,8 +290,14 @@ class RemoteDataSource {
     required String id,
     required String data,
   }) async {
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
     try {
       final update = await FirebaseFirestore.instance
+          .collection('app')
+          .doc(userId)
           .collection('finance')
           .doc(id)
           .update({'data': data});
@@ -173,13 +308,29 @@ class RemoteDataSource {
   }
 
   Future<void> removeBudgetDocuments({required String id}) {
-    return FirebaseFirestore.instance.collection('finance').doc(id).delete();
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
+    return FirebaseFirestore.instance
+        .collection('app')
+        .doc(userId)
+        .collection('finance')
+        .doc(id)
+        .delete();
   }
 
   Stream<QuerySnapshot<Map<String, dynamic>>> getAddSpendingsStream() {
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
     try {
-      final stream =
-          FirebaseFirestore.instance.collection('spendings').snapshots();
+      final stream = FirebaseFirestore.instance
+          .collection('app')
+          .doc(userId)
+          .collection('spendings')
+          .snapshots();
       return stream;
     } catch (error) {
       throw Exception(error.toString());
@@ -190,7 +341,15 @@ class RemoteDataSource {
     required String name,
     required String price,
   }) async {
-    await FirebaseFirestore.instance.collection('spendings').add(
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
+    await FirebaseFirestore.instance
+        .collection('app')
+        .doc(userId)
+        .collection('spendings')
+        .add(
       {
         'name': name,
         'outgoing': price,
@@ -199,13 +358,29 @@ class RemoteDataSource {
   }
 
   Future<void> removeSpendings({required String id}) {
-    return FirebaseFirestore.instance.collection('spendings').doc(id).delete();
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
+    return FirebaseFirestore.instance
+        .collection('app')
+        .doc(userId)
+        .collection('spendings')
+        .doc(id)
+        .delete();
   }
 
   Stream<QuerySnapshot<Map<String, dynamic>>> getAttractionStream() {
     try {
-      final stream =
-          FirebaseFirestore.instance.collection('attraction').snapshots();
+      final userId = FirebaseAuth.instance.currentUser?.uid;
+      if (userId == null) {
+        throw Exception('User is not logged in');
+      }
+      final stream = FirebaseFirestore.instance
+          .collection('app')
+          .doc(userId)
+          .collection('attraction')
+          .snapshots();
       return stream;
     } catch (error) {
       throw Exception(error.toString());
@@ -213,7 +388,15 @@ class RemoteDataSource {
   }
 
   Future<void> addAttraction({required String title}) async {
-    await FirebaseFirestore.instance.collection('attraction').add(
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
+    await FirebaseFirestore.instance
+        .collection('app')
+        .doc(userId)
+        .collection('attraction')
+        .add(
       {
         'title': title,
       },
@@ -221,6 +404,15 @@ class RemoteDataSource {
   }
 
   Future<void> removeAttraction({required String id}) {
-    return FirebaseFirestore.instance.collection('attraction').doc(id).delete();
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
+    return FirebaseFirestore.instance
+        .collection('app')
+        .doc(userId)
+        .collection('attraction')
+        .doc(id)
+        .delete();
   }
 }
